+++ plugin/src/main/java/com/stratio/cassandra/lucene/schema/mapping/Mapper.java
@@ -36,8 +36,6 @@ import java.util.function.Function;
 
 import static java.util.stream.Collectors.toList;
 
-import static java.util.stream.Collectors.toList;
-
 /**
  * Class for mapping between Cassandra's columns and Lucene documents.
  *
@@ -64,6 +62,8 @@ public abstract class Mapper {
 
     static final List<Class<?>> PRINTABLE_TYPES = Arrays.asList(
             String.class, Number.class, UUID.class, Boolean.class, InetAddress.class);
+
+    static final List<Class<?>> EMPTY_TYPE_LIST = Collections.emptyList();
 
     /** The store field in Lucene default option. */
     static final Store STORE = Store.NO;
@@ -92,6 +92,12 @@ public abstract class Mapper {
     /** The supported column value data types. */
     public final List<Class<?>> supportedTypes;
 
+    /** The explicitly excluded column value data types. */
+    public final List<Class<?>> excludedTypes;
+
+    /** If this mapper support collections. */
+    public final Boolean supportsCollections;
+
     /**
      * Builds a new {@link Mapper} supporting the specified types for indexing.
      *
@@ -101,13 +107,17 @@ public abstract class Mapper {
      * @param analyzer the name of the analyzer to be used
      * @param mappedColumns the names of the columns to be mapped
      * @param supportedTypes the supported column value data types
+     * @param excludedTypes the explicitly excluded value data types
+     * @param supportsCollections if this mapper supports collections
      */
     protected Mapper(String field,
                      Boolean docValues,
                      Boolean validated,
                      String analyzer,
                      List<String> mappedColumns,
-                     List<Class<?>> supportedTypes) {
+                     List<Class<?>> supportedTypes,
+                     List<Class<?>> excludedTypes,
+                     Boolean supportsCollections) {
         if (StringUtils.isBlank(field)) {
             throw new IndexException("Field name is required");
         }
@@ -118,6 +128,8 @@ public abstract class Mapper {
         this.mappedColumns = mappedColumns.stream().filter(Objects::nonNull).collect(toList()); // Remove nulls
         this.mappedCells = this.mappedColumns.stream().map(Column::parseCellName).collect(toList());
         this.supportedTypes = supportedTypes;
+        this.excludedTypes= excludedTypes;
+        this.supportsCollections = supportsCollections;
     }
 
     /**