+++ plugin/src/main/scala/com/stratio/cassandra/lucene/index/TokenLengthAnalyzer.scala
@@ -63,4 +63,3 @@ class TokenLengthAnalyzer(val analyzer: Analyzer)
     MoreObjects.toStringHelper(this).add("analyzer", analyzer).toString
   }
 }
-