+++ plugin/src/main/scala/com/stratio/cassandra/lucene/util/SchemaValidator.scala
@@ -23,8 +23,10 @@ import java.util.{Date, UUID}
 import com.stratio.cassandra.lucene.IndexException
 import com.stratio.cassandra.lucene.column.Column
 import com.stratio.cassandra.lucene.schema.Schema
+import com.typesafe.scalalogging.Logger
 import org.apache.cassandra.config.CFMetaData
 import org.apache.cassandra.db.marshal._
+import org.slf4j.LoggerFactory
 
 import scala.annotation.tailrec
 import scala.collection.JavaConverters._
@@ -35,14 +37,16 @@ import scala.collection.JavaConverters._
   */
 object SchemaValidator {
 
+  protected val logger: Logger = Logger(LoggerFactory.getLogger(getClass.getName))
+
   /** Validates the specified [[Schema]] against the specified [[CFMetaData]].
     *
-    * @param schema a schema
+    * @param schema   a schema
     * @param metadata a table metadata
     */
   def validate(schema: Schema, metadata: CFMetaData): Unit = {
     for (mapper <- schema.mappers.values.asScala; column <- mapper.mappedColumns.asScala) {
-      validate(metadata, column, mapper.field, mapper.supportedTypes.asScala.toList)
+      validate(metadata, column, mapper.field, mapper.supportedTypes.asScala.toList, mapper.excludedTypes.asScala.toList, mapper.supportsCollections)
     }
   }
 
@@ -50,7 +54,10 @@ object SchemaValidator {
       metadata: CFMetaData,
       column: String,
       field: String,
-      supportedTypes: List[Class[_]]) {
+      supportedTypes: List[Class[_]],
+      excludedTypes: List[Class[_]],
+      supportsCollections: Boolean) {
+
 
     val cellName = Column.parseCellName(column)
     val cellDefinition = metadata.getColumnDefinition(UTF8Type.instance.decompose(cellName))
@@ -62,8 +69,8 @@ object SchemaValidator {
       throw new IndexException("Lucene indexes are not allowed on static columns as '{}'", column)
     }
 
-    def checkSupported(t: AbstractType[_], mapper: String) {
-      if (!supports(t, supportedTypes)) {
+    def checkSupported(t: AbstractType[_], mapper: String, supportsCollections: Boolean) {
+      if (!supports(t, supportedTypes, excludedTypes, supportsCollections)) {
         throw new IndexException(
           "Type '{}' in column '{}' is not supported by mapper '{}'",
           t,
@@ -73,9 +80,10 @@ object SchemaValidator {
     }
 
     val cellType = cellDefinition.`type`
+    cellType.isCollection
     val udtNames = Column.parseUdtNames(column)
     if (udtNames.isEmpty) {
-      checkSupported(cellType, cellName)
+      checkSupported(cellType, cellName, supportsCollections)
     } else {
       var column = Column.apply(cellName)
       var currentType = cellType
@@ -84,7 +92,7 @@ object SchemaValidator {
         childType(currentType, udtNames(i)) match {
           case None => throw new IndexException(
             s"No column definition '${column.mapper}' for field '$field'")
-          case Some(n) if i == udtNames.indices.last => checkSupported(n, column.mapper)
+          case Some(n) if i == udtNames.indices.last => checkSupported(n, column.mapper, supportsCollections)
           case Some(n) => currentType = n
         }
       }
@@ -92,27 +100,34 @@ object SchemaValidator {
   }
 
   @tailrec
-  def childType(parent: AbstractType[_], child: String): Option[AbstractType[_]] = parent match {
-    case t: ReversedType[_] => childType(t.baseType, child)
-    case t: SetType[_] => childType(t.nameComparator, child)
-    case t: ListType[_] => childType(t.valueComparator, child)
-    case t: MapType[_, _] => childType(t.valueComparator, child)
-    case t: UserType =>
-      (0 until t.fieldNames.size).find(t.fieldNameAsString(_) == child).map(t.fieldType)
-    case t: TupleType => (0 until t.size).find(_.toString == child).map(t.`type`)
-    case _ => None
+  def childType(parent: AbstractType[_], child: String): Option[AbstractType[_]] = {
+    parent match {
+      case t: ReversedType[_] => childType(t.baseType, child)
+      case t: SetType[_] => childType(t.nameComparator, child)
+      case t: ListType[_] => childType(t.valueComparator, child)
+      case t: MapType[_, _] if child == Column.MAP_KEY_SUFFIX => Some(t.getKeysType)
+      case t: MapType[_, _] if child == Column.MAP_VALUE_SUFFIX => Some(t.valueComparator)
+      case t: MapType[_, _] => childType(t.valueComparator, child)
+      case t: UserType =>
+        (0 until t.fieldNames.size).find(t.fieldNameAsString(_) == child).map(t.fieldType)
+      case t: TupleType => (0 until t.size).find(_.toString == child).map(t.`type`)
+      case _ => None
+    }
   }
 
   @tailrec
   def supports(
       candidateType: AbstractType[_],
-      supportedTypes: Seq[Class[_]]): Boolean = candidateType match {
-    case t: ReversedType[_] => supports(t.baseType, supportedTypes)
-    case t: SetType[_] => supports(t.getElementsType, supportedTypes)
-    case t: ListType[_] => supports(t.getElementsType, supportedTypes)
-    case t: MapType[_, _] => supports(t.getValuesType, supportedTypes)
+      supportedTypes: Seq[Class[_]],
+      excludedTypes: List[Class[_]],
+      supportsCollections: Boolean): Boolean = candidateType match {
+    case t: ReversedType[_] => supports(t.baseType, supportedTypes, excludedTypes, supportsCollections)
+    case t: SetType[_] => if (supportsCollections) supports(t.getElementsType, supportedTypes, excludedTypes, supportsCollections) else false
+    case t: ListType[_] => if (supportsCollections)  supports(t.getElementsType, supportedTypes, excludedTypes, supportsCollections) else false
+    case t: MapType[_, _] =>  if (supportsCollections) supports(t.getValuesType, supportedTypes, excludedTypes, supportsCollections) else false
     case _ =>
       val native = nativeType(candidateType)
+      if (excludedTypes.contains(native)) return false
       supportedTypes.exists(_ isAssignableFrom native)
   }
 
@@ -133,5 +148,4 @@ object SchemaValidator {
     case _: InetAddressType => classOf[InetAddress]
     case _ => throw new IndexException(s"Unsupported Cassandra data type: ${validator.getClass}")
   }
-
 }