+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/varia/LargeFieldIT.java
@@ -18,6 +18,8 @@ package com.stratio.cassandra.lucene.testsAT.varia;
 import com.stratio.cassandra.lucene.testsAT.BaseIT;
 import com.stratio.cassandra.lucene.testsAT.util.CassandraUtils;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 
 import java.io.IOException;
 import java.util.Arrays;
@@ -28,6 +30,7 @@ import static com.stratio.cassandra.lucene.builder.Builder.*;
 /**
  * @author Andres de la Pena <adelapena@stratio.com>
  */
+@RunWith(JUnit4.class)
 public class LargeFieldIT extends BaseIT {
 
     @Test