+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/varia/MultiMappingIT.java
@@ -20,12 +20,15 @@ import com.stratio.cassandra.lucene.testsAT.util.CassandraUtils;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 
 import static com.stratio.cassandra.lucene.builder.Builder.*;
 
 /**
  * @author Andres de la Pena <adelapena@stratio.com>
  */
+@RunWith(JUnit4.class)
 public class MultiMappingIT extends BaseIT {
 
     private static CassandraUtils utils;
@@ -51,7 +54,7 @@ public class MultiMappingIT extends BaseIT {
 
     @AfterClass
     public static void after() {
-        utils.dropKeyspace();
+        CassandraUtils.dropKeyspaceIfNotNull(utils);
     }
 
     @Test