+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/varia/ReadStaticColumnsIT.java
@@ -21,12 +21,15 @@ import com.stratio.cassandra.lucene.testsAT.util.CassandraUtilsSelect;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.JUnit4;
 
 import static com.stratio.cassandra.lucene.builder.Builder.match;
 
 /**
  * @author Eduardo Alonso {@literal <eduardoalonso@stratio.com>}
  */
+@RunWith(JUnit4.class)
 public class ReadStaticColumnsIT extends BaseIT {
 
     private static CassandraUtils utils;
@@ -57,7 +60,7 @@ public class ReadStaticColumnsIT extends BaseIT {
 
     @AfterClass
     public static void after() {
-        utils.dropKeyspace();
+        CassandraUtils.dropKeyspaceIfNotNull(utils);
     }
 
     @Test