+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/varia/SortWithSkinnyRowsIT.java
@@ -53,7 +53,7 @@ public class SortWithSkinnyRowsIT extends BaseIT {
 
     @AfterClass
     public static void after() {
-        utils.dropIndex().dropTable().dropKeyspace();
+        CassandraUtils.dropKeyspaceIfNotNull(utils);
     }
 
     @Test